+++ student_time_back/src/main/java/com/example/server/auth/AuthenticationService.java
@@ -1,11 +1,11 @@
-package com.example.servingwebcontent.auth;
+package com.example.server.auth;
 
-import com.example.servingwebcontent.config.JwtService;
-import com.example.servingwebcontent.token.Token;
-import com.example.servingwebcontent.token.TokenRepository;
-import com.example.servingwebcontent.token.TokenType;
-import com.example.servingwebcontent.user.User;
-import com.example.servingwebcontent.user.UserRepository;
+import com.example.server.config.JwtService;
+import com.example.server.token.Token;
+import com.example.server.token.TokenRepository;
+import com.example.server.token.TokenType;
+import com.example.server.user.User;
+import com.example.server.user.UserRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -33,17 +33,13 @@ public class AuthenticationService {
         .password(passwordEncoder.encode(request.getPassword()))
         .role(request.getRole())
         .build();
-    var findUser = this.repository.findByEmail(request.getEmail())
-              .orElseThrow();
     var savedUser = repository.save(user);
     var jwtToken = jwtService.generateToken(user);
     var refreshToken = jwtService.generateRefreshToken(user);
-    var responseUser = user.getId() + user.getEmail() + user.getRole();
     saveUserToken(savedUser, jwtToken);
     return AuthenticationResponse.builder()
         .accessToken(jwtToken)
         .refreshToken(refreshToken)
-        .userResp(responseUser)
         .build();
   }
 
@@ -62,7 +58,7 @@ public class AuthenticationService {
     saveUserToken(user, jwtToken);
     return AuthenticationResponse.builder()
         .accessToken(jwtToken)
-            .refreshToken(refreshToken)
+        .refreshToken(refreshToken)
         .build();
   }
