+++ student_time_back/src/main/java/com/example/server/config/ApplicationConfig.java
@@ -1,4 +1,4 @@
-package com.example.servingwebcontent.config;
+package com.example.server.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -11,7 +11,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-import com.example.servingwebcontent.user.UserRepository;
+import com.example.server.user.UserRepository;
 
 import lombok.RequiredArgsConstructor;
