+++ student_time_back/src/main/java/com/example/server/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.example.servingwebcontent.config;
+package com.example.server.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
