+++ student_time_back/src/main/java/com/example/server/config/JwtService.java
@@ -1,4 +1,4 @@
-package com.example.servingwebcontent.config;
+package com.example.server.config;
 
 
 import io.jsonwebtoken.Claims;
