+++ student_time_back/src/main/java/com/example/server/config/LogoutService.java
@@ -1,6 +1,5 @@
-package com.example.servingwebcontent.config;
+package com.example.server.config;
 
-import com.example.servingwebcontent.token.TokenRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -8,6 +7,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.stereotype.Service;
+
+import com.example.server.token.TokenRepository;
 
 @Service
 @RequiredArgsConstructor
