+++ student_time_back/src/main/java/com/example/server/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.example.servingwebcontent.config;
+package com.example.server.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.enums.SecuritySchemeIn;
