+++ student_time_back/src/main/java/com/example/server/token/TokenRepository.java
@@ -1,4 +1,4 @@
-package com.example.servingwebcontent.token;
+package com.example.server.token;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 public interface TokenRepository extends JpaRepository<Token, Integer> {
+
   @Query(value = """
       select t from Token t inner join User u\s
       on t.user.id = u.id\s
