+++ student_time_back/src/main/java/com/example/server/token/TokenType.java
@@ -1,0 +1,5 @@
+package com.example.server.token;
+
+public enum TokenType {
+  BEARER
+}
