--- student_time_back/src/main/java/com/example/servingwebcontent/token/TokenType.java
@@ -1,5 +1,0 @@
-package com.example.servingwebcontent.token;
-
-public enum TokenType {
-  BEARER
-}
