+++ student_time_back/src/main/resources/application.yml
@@ -15,7 +15,7 @@ spring:
     show-sql: false
     properties:
       hibernate:
-        format_sql: true
+        "[format_sql]": true
     database: postgresql
     database-platform: org.hibernate.dialect.PostgreSQLDialect
